+++ hutool-cache/src/main/java/cn/hutool/cache/impl/AbstractCache.java
@@ -88,12 +88,10 @@ public abstract class AbstractCache<K, V> implements Cache<K, V> {
 		final MutableObj<K> mKey = MutableObj.of(key);
 
 		// issue#3618 对于替换的键值对，不做满队列检查和清除
-		if (cacheMap.containsKey(mKey)) {
-			CacheObj<K, V> oldObj = cacheMap.get(mKey);
-			if (oldObj != null) {
-				onRemove(oldObj.key, oldObj.obj);
-				cacheMap.remove(mKey);
-			}
+		final CacheObj<K, V> oldObj = cacheMap.get(mKey);
+		if (null != oldObj) {
+			onRemove(oldObj.key, oldObj.obj);
+			// 存在相同key，覆盖之
 			cacheMap.put(mKey, co);
 		} else {
 			if (isFull()) {